@@ -39,68 +39,66 @@
     "@graphql-codegen/cli": "5.0.7",
     "@nestjs/schematics": "11.0.5",
     "@nestjs/testing": "11.1.0",
-    "@ngneat/spectator": "^19.5.0",
-    "@nx/angular": "21.1.2",
-    "@nx/devkit": "21.1.2",
-    "@nx/eslint": "21.1.2",
-    "@nx/eslint-plugin": "21.1.2",
-    "@nx/jest": "21.1.2",
-    "@nx/js": "21.1.2",
-    "@nx/nest": "21.1.2",
-    "@nx/node": "21.1.2",
-    "@nx/playwright": "21.1.2",
-    "@nx/web": "21.1.2",
-    "@nx/webpack": "21.1.2",
-    "@nx/workspace": "21.1.2",
+    "@ngneat/spectator": "^20.0.0",
+    "@nx/angular": "21.2.1",
+    "@nx/devkit": "21.2.1",
+    "@nx/eslint": "21.2.1",
+    "@nx/eslint-plugin": "21.2.1",
+    "@nx/jest": "21.2.1",
+    "@nx/js": "21.2.1",
+    "@nx/nest": "21.2.1",
+    "@nx/node": "21.2.1",
+    "@nx/playwright": "21.2.1",
+    "@nx/web": "21.2.1",
+    "@nx/webpack": "21.2.1",
+    "@nx/workspace": "21.2.1",
     "@playwright/test": "^1.45.1",
-    "@schematics/angular": "19.2.9",
+    "@schematics/angular": "20.0.5",
     "@swc-node/register": "1.10.10",
     "@swc/core": "1.12.6",
-    "@swc/helpers": "~0.5.11",
-    "@swc/helpers": "~0.5.11",
+    "@swc/helpers": "~0.5.17",
     "@total-typescript/ts-reset": "^0.6.1",
     "@trivago/prettier-plugin-sort-imports": "^5.2.0",
     "@types/dompurify": "^3.2.0",
-    "@types/jest": "29.5.14",
-    "@types/node": "^22.10.2",
+    "@types/jest": "30.0.0",
+    "@types/node": "^24.0.10",
     "@types/trusted-types": "^2.0.7",
-    "@typescript-eslint/eslint-plugin": "8.25.0",
-    "@typescript-eslint/parser": "8.25.0",
-    "@typescript-eslint/utils": "8.25.0",
+    "@typescript-eslint/eslint-plugin": "8.35.1",
+    "@typescript-eslint/parser": "8.35.1",
+    "@typescript-eslint/utils": "8.35.1",
     "cobertura-merge": "^1.0.4",
-    "eslint": "8.57.1",
+    "eslint": "9.30.1",
     "eslint-config-prettier": "10.1.5",
-    "eslint-plugin-jest": "^28.10.0",
-    "eslint-plugin-playwright": "^2.1.0",
-    "eslint-plugin-rxjs": "^5.0.3",
+    "eslint-plugin-jest": "^29.0.1",
+    "eslint-plugin-playwright": "^2.2.0",
     "husky": "^9.1.7",
     "jest": "29.7.0",
-    "jest-environment-jsdom": "29.7.0",
+    "jest-environment-jsdom": "^29.7.0",
     "jest-environment-node": "^29.7.0",
-    "jest-preset-angular": "14.5.5",
+    "jest-preset-angular": "14.6.0",
     "lcov-result-merger": "^5.0.1",
     "mongo-migrate-ts": "^2.0.0",
-    "nx": "21.1.2",
+    "nx": "21.2.1",
     "playwright": "^1.49.1",
     "prettier": "3.6.1",
     "semantic-release": "^24.2.0",
     "ts-jest": "29.2.5",
     "ts-node": "10.9.2",
-    "typescript": "5.7.3",
+    "typescript": "5.8.3",
     "wait-on": "^8.0.1"
   },
   "dependencies": {
-    "@angular/animations": "19.2.10",
-    "@angular/build": "19.2.11",
-    "@angular/common": "19.2.10",
-    "@angular/compiler": "19.2.10",
-    "@angular/core": "19.2.10",
-    "@angular/forms": "19.2.10",
-    "@angular/platform-browser": "19.2.10",
-    "@angular/platform-browser-dynamic": "19.2.10",
-    "@angular/pwa": "19.2.11",
-    "@angular/router": "19.2.10",
-    "@angular/service-worker": "19.2.10",
+    "@angular/animations": "20.0.6",
+    "@angular/build": "20.0.5",
+    "@angular/common": "20.0.6",
+    "@angular/compiler": "20.0.6",
+    "@angular/core": "20.0.6",
+    "@angular/forms": "20.0.6",
+    "@angular/platform-browser": "20.0.6",
+    "@angular/platform-browser-dynamic": "20.0.6",
+    "@angular/pwa": "20.0.5",
+    "@angular/router": "20.0.6",
+    "@angular/service-worker": "20.0.6",
     "@apollo/client": "^3.13.8",
     "@automapper/classes": "^8.8.1",
     "@automapper/core": "^8.8.1",
