@@ -1,4 +1,3 @@
-import { CommonModule } from '@angular/common';
 import { ChangeDetectionStrategy, Component, input } from '@angular/core';
 import { NzTagModule } from 'ng-zorro-antd/tag';
 import { NzToolTipModule } from 'ng-zorro-antd/tooltip';
@@ -7,9 +6,9 @@ import { UnitUnion } from '@kordis/shared/model';
 
 @Component({
 	selector: 'krd-unit-chip',
-	imports: [CommonModule, NzTagModule, NzToolTipModule],
+	imports: [NzTagModule, NzToolTipModule],
 	template: `
-		<ng-container *ngIf="unit() as unit">
+		@if (unit(); as unit) {
 			<nz-tag>
 				<span nz-icon></span>
 				@switch (unit.__typename) {
@@ -23,7 +22,7 @@ import { UnitUnion } from '@kordis/shared/model';
 					}
 				}
 			</nz-tag>
-		</ng-container>
+		}
 	`,
 	styles: ``,
 	changeDetection: ChangeDetectionStrategy.OnPush,
